new features added
Random seed generation
Saving of settings to a settings.ini
mediator is saved within the PNG file 
the job is saved as a Json file
sorting of the queue and deleting of jobs is now in a separate tab 
settings can be modified and saved as default in the settings. Tab. 
Default settings can be restored to original default settings in the settings tab